# backend/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Distributed Job Scheduler API",
    "version": "1.0.0",
    "description": "API documentation for the Distributed Job Scheduler system"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateQueueDto": {
        "type": "object",
        "required": [
          "name",
          "projectId"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "concurrency": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "default": 5
          },
          "jobTimeoutMs": {
            "type": "integer",
            "default": 30000
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4,
            "default": 3
          },
          "retryPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "rateLimit": {
            "type": "integer"
          },
          "rateLimitWindow": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/workers/register": {
      "post": {
        "summary": "Register a new worker",
        "tags": [
          "Workers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "queueId": {
                    "type": "string"
                  },
                  "hostname": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Worker registered"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/workers/{id}/heartbeat": {
      "post": {
        "summary": "Record worker heartbeat",
        "tags": [
          "Workers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpuUsage": {
                    "type": "number"
                  },
                  "memoryUsage": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Heartbeat recorded"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/workers": {
      "get": {
        "summary": "List all workers",
        "tags": [
          "Workers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "queueId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of workers"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/queues": {
      "post": {
        "summary": "Create a new queue",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQueueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Queue created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List all queues",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of queues"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/queues/{id}": {
      "get": {
        "summary": "Get a queue by ID",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Queue not found"
          }
        }
      },
      "put": {
        "summary": "Update a queue",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQueueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queue updated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Queue not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a queue",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Queue deleted"
          },
          "404": {
            "description": "Queue not found"
          }
        }
      }
    },
    "/api/queues/{id}/pause": {
      "post": {
        "summary": "Pause a queue",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue paused"
          },
          "404": {
            "description": "Queue not found"
          }
        }
      }
    },
    "/api/queues/{id}/resume": {
      "post": {
        "summary": "Resume a paused queue",
        "tags": [
          "Queues"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue resumed"
          },
          "404": {
            "description": "Queue not found"
          }
        }
      }
    },
...
    "/api/dashboard/workers": {
      "get": {
        "summary": "Get worker statistics",
        "tags": [
          "Dashboard"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Worker stats"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}